@@ -9,10 +9,9 @@ import android.content.Context.ACTIVITY_SERVICE
 import android.content.Intent
 import androidx.core.text.isDigitsOnly
 import com.example.clock.data.model.Alarm
-import com.example.clock.util.Constants.dateTimeFormatter
-import com.example.clock.util.Constants.nextDay
+import com.example.clock.util.GlobalProperties.dateTimeFormatter
 import java.time.LocalDateTime
-import java.util.*
+import java.time.LocalTime
 
 fun String?.parseInt(): Int {
     return if (this == null || this.isEmpty()) 0 else this.toInt()
@@ -22,17 +21,14 @@ fun String.checkNumberPicker(maxNumber: Int): Boolean {
     return this.length <= 2 && this.isDigitsOnly() && this.parseInt() <= maxNumber
 }
 
+
 fun Alarm.checkDate(): String {
-    val currentTime = Calendar.getInstance()
-    val timeToMatch = Calendar.getInstance()
-    val currentDay = LocalDateTime.now()
-    timeToMatch[Calendar.HOUR_OF_DAY] = this.hour.parseInt()
-    timeToMatch[Calendar.MINUTE] = this.minute.parseInt()
+    val currentTime = LocalDateTime.now()
+    val alarmTime = LocalDateTime.of(currentTime.toLocalDate(), LocalTime.of(this.hour.parseInt(), this.minute.parseInt()))
 
     return when {
-        currentTime < timeToMatch -> "Today-${currentDay.format(dateTimeFormatter)}"
-        currentTime >= timeToMatch -> "Tomorrow-${nextDay.format(dateTimeFormatter)}"
-        else -> this.description
+        currentTime.isBefore(alarmTime) -> "Today-${currentTime.format(dateTimeFormatter)}"
+        else -> "Tomorrow-${alarmTime.plusDays(1).format(dateTimeFormatter)}"
     }
 }
 
@@ -49,7 +45,7 @@ fun Class<out BroadcastReceiver>?.setIntentAction(
         context,
         requestCode,
         broadcastIntent,
-        Constants.pendingIntentFlags,
+        GlobalProperties.pendingIntentFlags,
     )
 }
 
