@@ -4,23 +4,20 @@ import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
+import android.os.Handler
+import android.os.Looper
 import android.widget.Toast
 import com.example.clock.data.model.Alarm
 import com.example.clock.data.receiver.AlarmBroadcastReceiver
+import com.example.clock.data.receiver.DAYS_SELECTED
 import com.example.clock.data.receiver.HOUR
-import com.example.clock.data.receiver.IS_FRIDAY
-import com.example.clock.data.receiver.IS_MONDAY
 import com.example.clock.data.receiver.IS_RECURRING
-import com.example.clock.data.receiver.IS_SATURDAY
-import com.example.clock.data.receiver.IS_SUNDAY
-import com.example.clock.data.receiver.IS_THURSDAY
-import com.example.clock.data.receiver.IS_TUESDAY
-import com.example.clock.data.receiver.IS_WEDNESDAY
 import com.example.clock.data.receiver.MINUTE
 import com.example.clock.data.receiver.TITLE
-import com.example.clock.data.repository.AlarmRepository
-import com.example.clock.util.Constants.pendingIntentFlags
+import com.example.clock.util.GlobalProperties.pendingIntentFlags
 import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.flow.asFlow
+import kotlinx.coroutines.flow.buffer
 import java.util.Calendar
 import java.util.Random
 import javax.inject.Inject
@@ -29,81 +26,72 @@ import javax.inject.Singleton
 @Singleton
 class ScheduleAlarmManager @Inject constructor(
     @ApplicationContext private val applicationContext: Context,
-    private val alarmRepository: AlarmRepository,
 ) {
 
+    private val handler = Handler(Looper.getMainLooper())
+
     fun schedule(alarm: Alarm) {
-        val alarmManager =
-            applicationContext.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-
-        val intent = Intent(applicationContext, AlarmBroadcastReceiver::class.java).apply {
-            putExtra(IS_RECURRING, alarm.isRecurring); putExtra(IS_MONDAY, alarm.isMonday)
-            putExtra(IS_TUESDAY, alarm.isTuesday); putExtra(IS_WEDNESDAY, alarm.isWednesday)
-            putExtra(IS_THURSDAY, alarm.isThursday); putExtra(IS_FRIDAY, alarm.isFriday)
-            putExtra(IS_SATURDAY, alarm.isSaturday); putExtra(IS_SUNDAY, alarm.isSunday)
-            putExtra(TITLE, alarm.title); putExtra(HOUR, alarm.hour)
+        val alarmManager = applicationContext.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+        val alarmIntent = Intent(applicationContext, AlarmBroadcastReceiver::class.java).apply {
+            putExtra(IS_RECURRING, alarm.isRecurring)
             putExtra(MINUTE, alarm.minute)
+            putExtra(TITLE, alarm.title)
+            putExtra(HOUR, alarm.hour)
+            putExtra(DAYS_SELECTED, HashMap(alarm.daysSelected))
         }
-
         val alarmPendingIntent = PendingIntent.getBroadcast(
             applicationContext,
             alarm.id,
-            intent,
+            alarmIntent,
             pendingIntentFlags,
         )
-        val calendar: Calendar = Calendar.getInstance().apply {
-            timeInMillis = System.currentTimeMillis()
+        val calendar = Calendar.getInstance().apply {
             set(Calendar.HOUR_OF_DAY, alarm.hour.toInt())
             set(Calendar.MINUTE, alarm.minute.toInt())
             set(Calendar.SECOND, 0)
             set(Calendar.MILLISECOND, 0)
+            if (timeInMillis <= System.currentTimeMillis()) {
+                add(Calendar.DAY_OF_MONTH, 1)
+            }
+        }
+        val toastText = if (alarm.isRecurring) {
+            "Recurring Alarm ${alarm.title} scheduled for ${alarm.description} at ${alarm.hour}:${alarm.minute}"
+        } else {
+            "One Time Alarm ${alarm.title} scheduled for ${alarm.description.substringAfter('-')} at ${alarm.hour}:${alarm.minute}"
+        }
+        handler.post {
+            Toast.makeText(applicationContext, toastText, Toast.LENGTH_SHORT).show()
         }
 
-        if (calendar.timeInMillis <= System.currentTimeMillis()) {
-            calendar.set(Calendar.DAY_OF_MONTH, calendar.get(Calendar.DAY_OF_MONTH) + 1)
-        }
-
-        when (alarm.isRecurring) {
-            false -> {
-                val toastText = "One Time Alarm ${alarm.title} scheduled for" +
-                    " ${alarm.description.substringAfter('-')} at ${alarm.hour}:${alarm.minute}"
-
-                Toast.makeText(applicationContext, "${alarm.id}", Toast.LENGTH_LONG).show()
-                alarmManager.setExact(
-                    AlarmManager.RTC_WAKEUP,
-                    calendar.timeInMillis,
-                    alarmPendingIntent,
-                )
-            }
-            true -> {
-                val toastText = "Recurring Alarm ${alarm.title} scheduled for" +
-                    " ${alarm.description} at ${alarm.hour}:${alarm.minute}"
-
-                Toast.makeText(applicationContext, toastText, Toast.LENGTH_LONG).show()
-
-                val runDaily = (24 * 60 * 60 * 1000).toLong()
-                alarmManager.setRepeating(
-                    AlarmManager.RTC_WAKEUP,
-                    calendar.timeInMillis,
-                    runDaily,
-                    alarmPendingIntent,
-                )
-            }
+        if (alarm.isRecurring) {
+            alarmManager.setRepeating(
+                AlarmManager.RTC_WAKEUP,
+                calendar.timeInMillis,
+                AlarmManager.INTERVAL_DAY,
+                alarmPendingIntent,
+            )
+        } else {
+            alarmManager.setExactAndAllowWhileIdle(
+                AlarmManager.RTC_WAKEUP,
+                calendar.timeInMillis,
+                alarmPendingIntent,
+            )
         }
     }
 
     fun cancel(alarm: Alarm) {
-        val alarmManager =
-            applicationContext.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        val intent = Intent(applicationContext, AlarmBroadcastReceiver::class.java)
+        val alarmManager = applicationContext.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+        val alamIntent = Intent(applicationContext, AlarmBroadcastReceiver::class.java)
         val alarmPendingIntent = PendingIntent.getBroadcast(
             applicationContext,
             alarm.id,
-            intent,
+            alamIntent,
             pendingIntentFlags,
         )
         val toastText = "Alarm canceled for ${alarm.hour}:${alarm.minute}"
-        Toast.makeText(applicationContext, "${alarm.id}", Toast.LENGTH_SHORT).show()
+        handler.post {
+            Toast.makeText(applicationContext, toastText, Toast.LENGTH_SHORT).show()
+        }
 
         alarmManager.cancel(alarmPendingIntent)
     }
@@ -115,23 +103,18 @@ class ScheduleAlarmManager @Inject constructor(
 
         val alarm = Alarm(
             id = Random().nextInt(Integer.MAX_VALUE),
-            hour = calendar.get(Calendar.HOUR_OF_DAY).toString(),
-            minute = calendar.get(Calendar.MINUTE).toString(),
+            hour = String.format("%02d", calendar.get(Calendar.HOUR_OF_DAY)),
+            minute = String.format("%02d", calendar.get(Calendar.MINUTE)),
             title = "Snooze",
             isScheduled = true,
-            description = "Today",
         )
 
         schedule(alarm)
     }
 
-    fun cancelAlarms(alarms: List<Alarm>) {
-        for (alarm in alarms) {
-            if (alarm.isScheduled) {
-                cancel(alarm)
-            }
+    suspend fun clearScheduledAlarms(alarmsList: List<Alarm>) {
+        alarmsList.asFlow().buffer().collect {
+            cancel(it)
         }
     }
 }
-
-private const val TAG = "AlarmManager"
