@@ -3,14 +3,17 @@ package com.example.clock.data.model
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import com.example.clock.util.GlobalProperties.dateTimeFormatter
+import com.example.clock.util.GlobalProperties.nextDay
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 
 @Entity(
     tableName = "alarms_list_table",
     indices = [
         Index(
             value = [
-                "hour", "minute", "title", "description",
-                "isSunday", "isMonday", "isTuesday", "isWednesday", "isFriday", "isSaturday",
+                "hour", "minute", "title", "description", "daysSelectedJson",
             ],
             unique = true,
         ),
@@ -22,14 +25,25 @@ data class Alarm(
     var hour: String = "00",
     var minute: String = "00",
     var title: String = "",
-    var description: String = "",
+    var description: String = "Tomorrow-${nextDay.format(dateTimeFormatter)}",
     var isScheduled: Boolean = false,
     var isRecurring: Boolean = false,
-    var isSunday: Boolean = false,
-    var isMonday: Boolean = false,
-    var isTuesday: Boolean = false,
-    var isWednesday: Boolean = false,
-    var isThursday: Boolean = false,
-    var isFriday: Boolean = false,
-    var isSaturday: Boolean = false,
-)
+    var daysSelectedJson: String = Gson().toJson(
+        mapOf(
+            "Sun" to false,
+            "Mon" to false,
+            "Tue" to false,
+            "Wed" to false,
+            "Thu" to false,
+            "Fri" to false,
+            "Sat" to false,
+        ),
+    ),
+) {
+    val daysSelected: Map<String, Boolean>
+        get() = Gson().fromJson(daysSelectedJson, object : TypeToken<Map<String, Boolean>>() {}.type)
+
+    fun setDaysSelected(daysSelected: Map<String, Boolean>) {
+        this.daysSelectedJson = Gson().toJson(daysSelected)
+    }
+}
