@@ -1,16 +1,22 @@
 package com.example.clock.data.service
 
+import android.app.Service
 import android.content.Intent
-import androidx.lifecycle.LifecycleService
-import androidx.lifecycle.lifecycleScope
+import android.os.IBinder
 import com.example.clock.data.manager.ScheduleAlarmManager
 import com.example.clock.data.repository.AlarmRepository
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.flow.buffer
+import kotlinx.coroutines.flow.onCompletion
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class RescheduleAlarmsService : LifecycleService() {
+class RescheduleAlarmsService : Service() {
+
+    private val serviceScope = CoroutineScope(SupervisorJob())
 
     @Inject
     lateinit var scheduleAlarmManager: ScheduleAlarmManager
@@ -21,17 +27,20 @@ class RescheduleAlarmsService : LifecycleService() {
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         super.onStartCommand(intent, flags, startId)
 
-        lifecycleScope.launch {
-            alarmRepository.alarmsList.collect { alarms ->
-                for (alarm in alarms) {
-                    if (alarm.isScheduled) {
-                        scheduleAlarmManager.schedule(alarm)
+        serviceScope.launch {
+            alarmRepository.alarmsList.onCompletion {
+                stopForeground(STOP_FOREGROUND_REMOVE)
+            }
+                .buffer().collect { alarms ->
+                    for (alarm in alarms) {
+                        if (alarm.isScheduled) {
+                            scheduleAlarmManager.schedule(alarm)
+                        }
                     }
                 }
-            }
         }
         return START_STICKY
     }
-}
 
-private const val TAG = "RescheduleAlarmsService"
+    override fun onBind(intent: Intent): IBinder? = null
+}
