@@ -1,6 +1,5 @@
 package com.example.clock.ui.alarm
 
-import android.util.Log
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
@@ -10,37 +9,34 @@ import androidx.lifecycle.viewModelScope
 import com.example.clock.data.manager.ScheduleAlarmManager
 import com.example.clock.data.model.Alarm
 import com.example.clock.data.repository.AlarmRepository
-import com.example.clock.util.Constants.alarmDefaultValue
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-private const val TAG = "AlarmsListViewModel"
-
 @HiltViewModel
 class AlarmViewModel @Inject constructor(
     private val alarmRepository: AlarmRepository,
     private val scheduleAlarmManager: ScheduleAlarmManager,
-) : ViewModel() {
+) : ViewModel(), AlarmActions {
 
     val alarmsListState = alarmRepository.alarmsList.asLiveData()
-    var createAlarmState by mutableStateOf(alarmDefaultValue)
+
+    var alarmCreationState by mutableStateOf(Alarm())
         private set
 
-    fun changeCreateAlarmState(alarm: Alarm) {
-        createAlarmState = alarm
+    override fun updateAlarmCreationState(alarm: Alarm) {
+        alarm.setDaysSelected(alarm.daysSelected)
+        alarmCreationState = alarm
     }
 
-    fun update(alarm: Alarm) = viewModelScope.launch {
-        Log.e(TAG, "description: ${alarm.description}")
-        alarmRepository.update(alarm)
-    }
-
-    fun onScheduledChange(alarm: Alarm) {
+    override fun update(alarm: Alarm) {
         viewModelScope.launch {
             listOf(
-                async { update(alarm) },
+                async {
+                    alarm.setDaysSelected(alarm.daysSelected)
+                    alarmRepository.update(alarm)
+                },
                 async {
                     if (alarm.isScheduled) {
                         scheduleAlarmManager.schedule(alarm)
@@ -52,7 +48,7 @@ class AlarmViewModel @Inject constructor(
         }
     }
 
-    fun remove(alarm: Alarm) {
+    override fun remove(alarm: Alarm) {
         viewModelScope.launch {
             listOf(
                 async { alarmRepository.delete(alarm) },
@@ -65,35 +61,36 @@ class AlarmViewModel @Inject constructor(
         }
     }
 
-    fun saveAlarm() {
+    override fun save() {
         viewModelScope.launch {
-            val lastAutoGeneratedId = alarmRepository.getLastAutoId()
-            val alarm = alarmRepository.getAlarmById(createAlarmState.id)
+            val lastId = alarmRepository.getLastId()
+            val alarm = alarmRepository.getAlarmById(alarmCreationState.id)
 
-            if (!createAlarmState.isScheduled) {
-                createAlarmState.isScheduled = true
+            if (!alarmCreationState.isScheduled) {
+                alarmCreationState.isScheduled = true
             }
 
             listOf(
                 async {
-                    if (alarm?.id == createAlarmState.id) {
-                        update(createAlarmState)
+                    if (alarm?.id == alarmCreationState.id) {
+                        update(alarmCreationState)
                     } else {
-                        createAlarmState.id = lastAutoGeneratedId?.plus(1) ?: 1
-                        alarmRepository.insert(createAlarmState)
+                        alarmCreationState.id = lastId?.plus(1) ?: 1
+                        alarmCreationState.setDaysSelected(alarmCreationState.daysSelected)
+                        alarmRepository.insert(alarmCreationState)
                     }
                 },
-                async { scheduleAlarmManager.schedule(createAlarmState) },
+                async { scheduleAlarmManager.schedule(alarmCreationState) },
             )
         }
     }
 
-    fun clearAlarmsList() {
+    override fun clear() {
         viewModelScope.launch {
             alarmsListState.value?.let {
                 listOf(
                     async { alarmRepository.clear() },
-                    async { scheduleAlarmManager.cancelAlarms(alarms = it) },
+                    async { scheduleAlarmManager.clearScheduledAlarms(it) },
                 )
             }
         }
